+++ WSProductos1001.Domain/Features/Vehicle/VehicleService.cs
@@ -1,20 +1,24 @@
 ﻿using FluentValidation;
+using WSProductos1001.Domain.Features.RequestCredit;
 using WSProductos1001.Domain.Repository;
 using WSProductos1001.Domain.Services;
 using WSProductos1001.Entities;
 
 namespace WSProductos1001.Domain.Features.Vehicle
 {
-    internal class VehicleService : IVehicleService
+    public class VehicleService : IVehicleService
     {
         private readonly IVehicleRepository _vehicleRepository;
         private readonly IBrandRepository _brandRepository;
         private readonly IValidator<EVehicle> _validator;
-        public VehicleService(IVehicleRepository vehicleRepository, IValidator<EVehicle> validator, IBrandRepository brandRepository)
+        private readonly IRequestCreditRepository _requestCreditRepository;
+        public VehicleService(IVehicleRepository vehicleRepository, IValidator<EVehicle> validator,
+            IBrandRepository brandRepository, IRequestCreditRepository requestCreditRepository)
         {
             _vehicleRepository = vehicleRepository;
             _validator = validator;
             _brandRepository = brandRepository;
+            _requestCreditRepository = requestCreditRepository;
         }
         
 
@@ -34,6 +38,8 @@ namespace WSProductos1001.Domain.Features.Vehicle
         {
             var vehicle = await _vehicleRepository.GetByIdAsync(id);
             if (vehicle == null) throw new Exceptions.NotFoundException(nameof(EVehicle), id);
+            var vehcileRequest = await _requestCreditRepository.GetRequestCreditByVehicleRegistry(id, (int) CreditStatus.Registry);
+            if (vehcileRequest != null) throw new Exceptions.BadRequestException("El vehiculo no puede ser eliminado porque tiene un credito pendiente");
             await _vehicleRepository.DeleteAsync(vehicle);
             
         }
